app/web_gui_service/requirements.txt
flask
flask_socketio
paho.mqtt
requests